Simplify doc
